requirements.txt
Werkzeug==2.3.7
Flask==2.3.2
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.3
email-validator==1.3.1
gunicorn==20.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-flask==1.3.0
pytest-mock==3.11.1
selenium==4.15.2

# Email Notifications
Flask-Mail==0.9.1

# Admin Dashboard
Flask-Admin==1.6.1

# Payments (Razorpay)
razorpay==1.4.1

# Real-time Features (WebSockets)
Flask-SocketIO==5.3.6
python-socketio==5.8.0
eventlet==0.33.3

# API Authentication (JWT)
PyJWT==2.8.0
flask-jwt-extended==4.5.3

# File Uploads
Flask-Uploads==0.2.1
Pillow==10.0.1

# Charts/Analytics
matplotlib==3.7.2
plotly==5.15.0

# Additional utilities
requests==2.31.0
bcrypt==4.0.1
